implement custom exception

+++ src/main/java/org/example/bankingsystem/services/AccountService.java
@@ -2,11 +2,12 @@ package org.example.bankingsystem.services;
 
 import org.example.bankingsystem.entities.Account;
 import org.example.bankingsystem.entities.User;
+import org.example.bankingsystem.exceptions.AccountNotFoundException;
+import org.example.bankingsystem.exceptions.UserNotFoundException;
 import org.example.bankingsystem.repository.AccountRepository;
 import org.example.bankingsystem.repository.UserRepository;
 import org.example.bankingsystem.requests.AccountRequest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -20,11 +21,24 @@ public class AccountService {
     @Autowired
     private UserRepository userRepository;
     //method to create account number
-    public Account createAccount(AccountRequest request){
+    public Account createAccount(AccountRequest request) {
+
         Account account = new Account();
         account.setAccountType(request.getAccountType());
-//        account.setBalance(request.getBalance());
-        account.setUser(request.getUser());
+        User user = userRepository.findById(request.getUser().getUserId())
+                .orElseThrow(() ->
+                        new UserNotFoundException("User not found with Id : " + request.getUser().getUserId()));
+
+//        Optional<User> optionalUser =
+//                userRepository.findById(request.getUser().getUserId());
+//
+//        if (optionalUser.isEmpty()) {
+//            throw new UserNotFoundException(
+//                    "User not found with Id : " + request.getUser().getUserId());
+//        }
+//        User user = optionalUser.get();
+        account.setUser(user);
+
         return accountRepository.save(account);
     }
     //getting balance of the account
@@ -34,16 +48,24 @@ public class AccountService {
             Account a = account.get();
             return a.getBalance();
         }
-        return null;
+        throw new AccountNotFoundException("Account not found with account number : "+accNo);
     }
     //getting one or more account number of user
     public List<Account> showAccounts(Integer userId) {
         if (!userRepository.existsById(userId)) {
-            return null;
+            throw new UserNotFoundException("User not found with Id : "+userId);
         }
-        return accountRepository.findByUserUserId(userId);
+        List<Account> accounts = accountRepository.findByUserUserId(userId);
+        if (accounts.isEmpty()){
+            throw new AccountNotFoundException("No Accounts found for User id : "+userId);
+        }
+        return accounts;
     }
     public void deleteAccount(long accountNo){
+        if (!accountRepository.existsById(accountNo)) {
+            throw new AccountNotFoundException("Account not found with account number : "+accountNo);
+
+        }
         accountRepository.deleteById(accountNo);
     }
